+++ frontend/public/changelog.md
@@ -14,6 +14,7 @@
 - Memisahkan jenis entri menjadi bentuk (dasar, turunan, gabungan) dan ekspresi (idiom, peribahasa)
 - Memisahkan kelas kata menjadi kelas kata dan unsur terikat
 - Memindahkan data makna.kelas_kata unsur terikat ke entri.jenis
+- Menambahkan fitur komentar pengguna di halaman kamus
 
 ## 2026-02-16 Senin